+++ tests/e2e-tests/src/orchestration.test.ts
@@ -1,12 +1,7 @@
-import path from 'path';
-import { fileURLToPath } from 'url';
-import dotenv from 'dotenv';
+import { loadEnv } from './utils/load-env.js';
 import { OrchestrationClient } from '@sap-ai-sdk/orchestration';
 
-// Pick .env file from e2e root directory
-const __filename = fileURLToPath(import.meta.url);
-const __dirname = path.dirname(__filename);
-dotenv.config({ path: path.resolve(__dirname, '../.env') });
+loadEnv();
 
 describe('orchestration', () => {
   it('should complete a chat', async () => {
